requirements.txt
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.7.0.72
matplotlib==3.7.1
PyWavelets==1.4.1
scikit-learn==1.3.0 